Command all-or-nothing
Command execution was simplified, not allowing to save events to persistence storage
in the middle of command execution. Now command result is a set of events, or nothing.
In case of exception in aggregate non of produced events will be saved, and aggregate will
be restored from persistence storage to process other commands.
Aggregate loose method Produce, only Emit is available